+++ tensorcircuit/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __author__ = "TensorCircuit Authors"
 __creator__ = "refraction-ray"