+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Add support of multiple tensor inputs for `tc.TorchLayer`
 
+- Both gate index and ``expectation_ps`` now support negative qubit index, eg. `c.H(-1)` with the same meaning as numpy indexing
+
 ### Changed
 
 - Change pytest xdist option in check_all.sh to `-n auto`